moved derived scatter class to plot.hpp
